pyproject.toml
[project]
name = "my-trading-prj"
version = "0.1.0"
description = ""
authors = [
    {name = "ArtemKalchenkoGit",email = "artem.kalchenko2004@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10 <4.0.0"
dependencies = [
    "langchain (>=1.2.0,<2.0.0)",
    "langgraph (>=1.0.5,<2.0.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "langchainhub (>=0.1.21,<0.2.0)",
    "chromadb (>=1.3.7,<2.0.0)",
    "tavily-python (>=0.7.17,<0.8.0)",
    "black (>=25.12.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langchain-ollama (>=1.0.1,<2.0.0)",
    "langchain-tavily (>=0.2.15,<0.3.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "ta (>=0.11.0,<0.12.0)",
    "yfinance (>=0.2.66,<0.3.0)",
    "datetime (>=6.0,<7.0)",
    "langchain-text-splitters (>=1.1.0,<2.0.0)",
    "langchain-chroma (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
